add update assert to file

+++ binding/src/main/java/com/lifecycle/binding/server/LocalServer.kt
@@ -9,16 +9,24 @@ import com.lifecycle.binding.util.assertToFile
 import java.io.File
 
 
-class LocalServer(private val json:String = "json",private val port:Int = 8888) : HttpServerRequestCallback {
-
-    private val path by lazy { AppLifecycle.application.getExternalFilesDir(null)!!
-            .also {
-                    it.deleteRecursively()
-                    assertToFile(AppLifecycle.application, json, it)
-            }
-    }
+class LocalServer(
+    private val json: String = "json",
+    private val port: Int = 8888,
+    private val update: Boolean
+) : HttpServerRequestCallback {
 
     private val server = AsyncHttpServer()
+    private val path by lazy {
+        AppLifecycle.application.getExternalFilesDir(null)!!
+            .also {
+                if (update) {
+                    it.deleteRecursively()
+                    assertToFile(AppLifecycle.application, json, it)
+                } else if (!File(it, json).isDirectory){
+                    assertToFile(AppLifecycle.application, json, it)
+                }
+            }
+    }
 
     fun start() {
         server.addAction("OPTIONS", "[\\d\\D]*", this)
@@ -28,10 +36,11 @@ class LocalServer(private val json:String = "json",private val port:Int = 8888) 
     }
 
     override fun onRequest(request: AsyncHttpServerRequest, response: AsyncHttpServerResponse) {
-        File(path,  json+request.path).also { if (it.exists()) response.sendFile(it) }
+        File(path, json + request.path).also { if (it.exists()) response.sendFile(it) }
     }
 
     fun stop() {
         server.stop()
     }
+
 }
